curl_project/backend/requirements.txt
# ====================
# Core Django Dependencies
# ====================
Django>=5.0,<6.0
python-dotenv>=1.0.1
sqlparse>=0.5.0
psycopg2-binary>=2.9.9
django-widget-tweaks>=1.5.0
Pillow>=10.0.0

# ====================
# Authentication & User Management
# ====================
djangorestframework>=3.16.0
dj-rest-auth>=7.0.1
django-allauth>=65.9.0
djangorestframework-simplejwt>=5.5.0
django-guest-user>=0.5.5

# ====================
# API Documentation
# ====================
drf-yasg>=1.21.0

# ====================
# Middleware & Headers
# ====================
django-cors-headers>=4.3.1

# ====================
# Debugging & Dev Utilities
# ====================
django-debug-toolbar>=4.2.0

# ====================
# Geolocation & User Data
# ====================
geoip2>=4.8.0
maxminddb>=2.6.1
IP2Location>=8.10.2
ip2geotools>=0.1.6
user-agents>=2.2.0
requests>=2.31.0

# ====================
# Code Quality & Testing Tools
# ====================
black>=23.3.0
isort>=5.13.2
pylint>=3.2.0
pytest>=7.2.0
